add option to supress recompilation (--no-force,-n)
in the past commit I add --force to the `cargo run` command. but now I
add tests for many of library, so recompiling of all tests takes longer
and longer time. especially, even for just checking the 'current status'
to decide what to do next, it always recompile all of tests.  this kind
of checking doesn't have to be as accurate as the final result, so I
added --no-force option (or -n) to get a rough result quickly.